Added printer drivers for CUPS - Gutenprint
Printer drivers for CUPS
This package includes a CUPS driver based on Gutenprint.
The CUPS drivers contain all of the files needed to support photo-quality printing on any printer supported by Gutenprint.
Gutenprint is the print facility for the GIMP, and in addition a suite of drivers that may be used with common UNIX spooling systems using GhostScript or CUPS. These drivers provide printing quality for UNIX/Linux on a par with proprietary vendor-supplied drivers in many cases, and can be used for many of the most demanding printing tasks. Gutenprint was formerly known as Gimp-Print.